+++ ElvUI_LuckyoneUI/Core/Tags.lua
@@ -31,6 +31,12 @@ E:AddTag('luckyone:health:percent', 'UNIT_HEALTH UNIT_MAXHEALTH', function(unit)
 	return E:GetFormattedText('PERCENT', currentHealth, maxHealth, percent == 100 and 0 or percent < 10 and 2 or 1, nil)
 end)
 
+-- Display percentage mana with 0 decimals
+E:AddTag('luckyone:mana:percent', 'UNIT_POWER_FREQUENT UNIT_MAXPOWER UNIT_DISPLAYPOWER', function(unit)
+	local min = UnitPower(unit, Enum.PowerType.Mana)
+	return E:GetFormattedText('PERCENT', min, UnitPowerMax(unit, Enum.PowerType.Mana), 0, nil)
+end)
+
 -- Display mana (current) if the unit is flagged healer (Retail only)
 E:AddTag('luckyone:healermana:current', 'UNIT_POWER_FREQUENT UNIT_MAXPOWER', function(unit)
 	local role = UnitGroupRolesAssigned(unit)
@@ -42,13 +48,9 @@ end, not E.Retail)
 -- Display mana (percent) if the unit is flagged healer (Retail only)
 E:AddTag('luckyone:healermana:percent', 'UNIT_MAXPOWER UNIT_POWER_FREQUENT', function(unit)
 	local role = UnitGroupRolesAssigned(unit)
+	local min = UnitPower(unit, Enum.PowerType.Mana)
 	if role == 'HEALER' then
-		local powerMax = UnitPowerMax(unit, Enum.PowerType.Mana)
-		if powerMax == 0 then
-			return 0
-		else
-			return floor(UnitPower(unit, Enum.PowerType.Mana) / powerMax * 100 + 0.5)
-		end
+		return E:GetFormattedText('PERCENT', min, UnitPowerMax(unit, Enum.PowerType.Mana), 0, nil)
 	end
 end, not E.Retail)
 
@@ -59,3 +61,4 @@ E:AddTagInfo('luckyone:healermana:percent', Private.Name, L["Displays the unit's
 -- Descriptions for Available Tags
 E:AddTagInfo('luckyone:classification', Private.Name, L["Displays the unit's classification (e.g 'Elite' and 'Rare') but without 'Affix'"])
 E:AddTagInfo('luckyone:health:percent', Private.Name, L["Displays percentage health with 1 decimal below 100%, 2 decimals below 10% and hides decimals at 100%"])
+E:AddTagInfo('luckyone:mana:percent', Private.Name, L["Displays percentage mana without decimals"])