+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+### Version 2.41 [18-09-2023]
+**All WoW Versions:**  
+Added new custom tag luckyone:mana:percent for mana without decimals  
+Party frames in the healing layout are back to vertical style  
+Removed Quartz profiles  
+Swapped default positions of Micro and Stance bar for fresh installs  
+Updated default position of VehicleSeatMover slightly  
+Updated default postion of DurabilityFrameMover slightly  
+**Retail:**  
+Updated OmniCD profile to match new party healing layout  
+**Wrath Classic:**  
+Match default position of Boss Frames with Retail  
+___
 ### Version 2.40 [06-09-2023]
 **All WoW Versions:**  
 ElvUI 13.40+ compatibility